Fix overflow issues

+++ src/main/java/seedu/address/model/person/Person.java
@@ -234,7 +234,7 @@ public record Person(
          * @param salaryInCents The salary to set.
          * @return This builder object.
          */
-        public Builder withSalaryInCents(int salaryInCents) {
+        public Builder withSalaryInCents(long salaryInCents) {
             this.salary = new Salary(salaryInCents / 100.0);
             return this;
         }
